Add continuous integration
Add continuous integration (CI) with GitHub Actions. It runs the unit tests and builds a debug APK. It does not block bad commits (gated commits) but you will at least be able to see if a pull request breaks anything before pulling it.

I used the Github-Android-Action action from the Marketplace [[1]] as a starting point and modified it.

[1]: https://github.com/marketplace/actions/android-github-action