Fix hottest time of day
The hottest time of day is usually around 15:00. The code assumed it was after 10:00 and before 14:00. This change adjusts the range to after 13:00 and before 17:00. This makes the daily high for the course of the week much more accurate.

It is still a hack because there might be a rain storm that comes through during those times. The TODO comment that was already present suggested to find the maximum instead, which is the correct solution.